+++ lib/draggable_menu.dart
@@ -1,9 +1,14 @@
 library draggable_menu;
 
+// Menu Related
 export 'src/draggable_menu/menu/menu.dart';
-export 'src/draggable_menu/menu/enums/status.dart';
-export 'src/draggable_menu/menu/ui.dart';
-export 'src/draggable_menu/utils/scrollable_manager/scrollable_manager.dart';
-export 'src/draggable_menu/menu/custom_draggable_menu.dart';
 export 'src/draggable_menu/menu/draggable_menu_level.dart';
+export 'src/draggable_menu/menu/enums/status.dart';
+
+// UI Related
+export 'src/draggable_menu/menu/ui.dart';
+export 'src/draggable_menu/menu/custom_draggable_menu.dart';
+
+// Usage Related
+export 'src/draggable_menu/utils/scrollable_manager/scrollable_manager.dart';
 export 'src/draggable_menu/menu/controller.dart';